backend/requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langgraph>=1.0.0
langchain-chroma>=1.0.0
langchain-core>=1.0.0
chromadb>=1.0.0
pypdf==3.17.4
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.9.0
openai>=1.12.0
sentence-transformers==2.2.2
numpy>=1.26.3
aiofiles==23.2.1
python-multipart==0.0.6

